serial number

+++ Inventory.DataObjects/EDM/Item.cs
@@ -22,7 +22,7 @@ namespace Inventory.DataObjects.EDM
         public string LocationInStock { get; set; }
         public Nullable<System.DateTime> ExpiryDate { get; set; }
         public int UnitID { get; set; }
-        public Nullable<int> Quantity { get; set; }
+        public Nullable<decimal> Quantity { get; set; }
         public Nullable<int> UnitAmount { get; set; }
         public Nullable<decimal> Price { get; set; }
         public Nullable<decimal> MaintenancePrice { get; set; }
